autotest extended: optional parallel, fast method find, save random construction parameters

+++ tsl2.nano.autotest/src/main/java/de/tsl2/nano/autotest/creator/AutoTestGenerator.java
@@ -2,26 +2,45 @@ package de.tsl2.nano.autotest.creator;
 
 import static de.tsl2.nano.autotest.creator.AFunctionCaller.def;
 import static de.tsl2.nano.autotest.creator.AFunctionTester.PREF_PROPS;
+import static java.nio.file.StandardOpenOption.*;
 
+import java.io.BufferedWriter;
 import java.io.File;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.lang.Thread.UncaughtExceptionHandler;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.nio.file.Files;
+import java.nio.file.OpenOption;
+import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Date;
 import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Scanner;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Stream;
 
+import de.tsl2.nano.core.ManagedException;
 import de.tsl2.nano.core.cls.BeanClass;
 import de.tsl2.nano.core.cls.ClassFinder;
 import de.tsl2.nano.core.cls.PrimitiveUtil;
+import de.tsl2.nano.core.exception.ExceptionHandler;
 import de.tsl2.nano.core.log.LogFactory;
 import de.tsl2.nano.core.util.ConcurrentUtil;
 import de.tsl2.nano.core.util.DateUtil;
 import de.tsl2.nano.core.util.FileUtil;
 import de.tsl2.nano.core.util.FormatUtil;
+import de.tsl2.nano.core.util.NumberUtil;
 import de.tsl2.nano.core.util.ObjectUtil;
 import de.tsl2.nano.core.util.StringUtil;
 import de.tsl2.nano.core.util.Util;
@@ -55,6 +74,8 @@ public class AutoTestGenerator {
 	static int filter_unsuccessful = 0;
 	static int filter_nullresults = 0;
 	static int filter_complextypes = 0;
+	private static BufferedWriter filteredFunctionWriter;
+	private static ExceptionHandler uncaughtExceptionHandler = new ExceptionHandler();
 	
 	public static void main(String[] args) {
 		List<Method> methods = ClassFinder.self().find(def("filter", "") , Method.class, def("modifier", -1), null);
@@ -67,8 +88,10 @@ public class AutoTestGenerator {
 				+ "\n\tuser.name              : " + System.getProperty("user.name")
 				+ "\n\tstart time             : " + DateUtil.getFormattedDateTime(new Date())
 				+ "\n\tfilename pattern       : " + fileName
+				+ "\n\tfast.classscan         : " + def("fast.classscan", true)
 				+ "\n\tclean                  : " + def("clean", false)
 				+ "\n\tduplication            : " + def("duplication", 10)
+				+ "\n\tparallel               : " + def("parallel", false)
 				+ "\n\tfilter                 : " + def("filter", "")
 				+ "\n\tmodifier               : " + def("modifier", -1)
 				+ "\n\tfilter.test            : " + def("filter.test", ".*(Test|IT)")
@@ -82,26 +105,54 @@ public class AutoTestGenerator {
 	}
 	@SuppressWarnings("rawtypes")
 	public static Collection<? extends AFunctionTester> createExpectationTesters() {
-		printStartParameters();
-		FileUtil.delete(fileName + "filtered.txt");
-		int duplication = def("duplication", 10);
-		List<Method> methods = ClassFinder.self().find(def("filter", "") , Method.class, def("modifier", -1), null);
-		filterExcludes(methods);
-		filterTestClasses(methods);
-		Collection<AFunctionTester> testers = new LinkedList<>();
-		boolean fileexists = true;
-		for (int i=0; i<duplication; i++) {
-			if (!AutoTestGenerator.getFile(i).exists() || def("clean", false)) {
-				fileexists = false;
-				AutoTestGenerator.generateExpectations(i, methods);
-			}
-			if (fileexists || count > 0)
-				testers.addAll(AutoTestGenerator.readExpectations(i));
+		try {
+			printStartParameters();
+			FileUtil.delete(fileName + "initialization-error.txt");
+			prepareFilteredWriter();
+			int duplication = def("duplication", 10);
+			List<Method> methods;
+			if (def("fast.classscan", true))
+				methods = ClassFinder.self().findMethods(def("filter", ""), def("modifier", -1), null);
+			else
+				methods = ClassFinder.self().find(def("filter", ""), Method.class, def("modifier", -1), null);
+			filterExcludes(methods);
+			filterTestClasses(methods);
+			
+			Collection<AFunctionTester> testers = Collections.synchronizedList(new LinkedList<>());
+			ArrayList<Integer> dupList = NumberUtil.numbers(duplication);
+			AtomicBoolean fileexists = new AtomicBoolean(true);
+			Util.stream(dupList, def("parallel", false)).forEach( i -> 
+			{
+				Thread.currentThread().setUncaughtExceptionHandler(uncaughtExceptionHandler );
+				if (!AutoTestGenerator.getFile(i).exists() || def("clean", false)) {
+					fileexists.set(false);
+					AutoTestGenerator.generateExpectations(i, methods);
+				}
+				if (fileexists.get() || count > 0)
+					testers.addAll(AutoTestGenerator.readExpectations(i));
+			});
+			if (def("filter.unsuccessful", true))
+				load_unsuccessful = FunctionCheck.filterFailingTest(testers);
+			printStatistics(duplication +1, testers);
+			return testers;
+		} catch (Exception e) {
+			ManagedException.writeError(e, fileName + "initialization-error.txt");
+			ConcurrentUtil.sleep(3000);
+			System.out.println("JUNIT TEST PARAMETERS FAILING:");
+			e.printStackTrace();
+			ManagedException.forward(e);
+			return null;
+		} finally {
+			Util.trY(() -> filteredFunctionWriter.close());
+//			if (uncaughtExceptionHandler.hasExceptions())
+//				throw new IllegalStateException(uncaughtExceptionHandler.toString());
 		}
-		if (def("filter.unsuccessful", true))
-			load_unsuccessful = FunctionCheck.filterFailingTest(testers);
-		printStatistics(duplication +1, testers);
-		return testers;
+	}
+	private static void prepareFilteredWriter() throws IOException {
+		if (!AutoTestGenerator.getFile(0).exists() || def("clean", false))
+			FileUtil.delete(fileName + "filtered.txt");
+		File userDirFile = FileUtil.userDirFile(fileName + "filtered.txt");
+		filteredFunctionWriter = Files.newBufferedWriter(Paths.get(userDirFile.getPath()), CREATE, WRITE, APPEND);
 	}
 
 	private static void filterExcludes(List<Method> methods) {
@@ -112,28 +163,34 @@ public class AutoTestGenerator {
 	}
 	static void generateExpectations(int iteration, List<Method> methods) {
 		LogFactory.setPrintToConsole(false);
-		count = 0;
-		methods_loaded = methods.size();
-		String p = "\n" + StringUtil.fixString(79, '~') + "\n";
-		AFunctionCaller.log(p + "calling " + methods.size() + " methods to create expectations -> " + getFile(iteration) + p);
-
-		methods.forEach(m -> writeExpectation(new AFunctionCaller(iteration, m)));
-		
-		ConcurrentUtil.sleep(500);
-		if (count > 0)
-			AFunctionCaller.log(new String(FileUtil.getFileBytes(getFile(iteration).getPath(), null)));
-		AFunctionCaller.log(p + count + " expectations written into '" + getFile(iteration) + p);
+		BufferedWriter writer = Util.trY(() -> Files.newBufferedWriter(Paths.get(getFile(iteration).getPath()), CREATE, WRITE, APPEND));
+		try {
+			count = 0;
+			methods_loaded = methods.size();
+			String p = "\n" + StringUtil.fixString(79, '~') + "\n";
+			AFunctionCaller.log(p + "calling " + methods.size() + " methods to create expectations -> " + getFile(iteration) + p);
+	
+			Util.stream(methods, def("parallel", false)).forEach(m -> writeExpectation(new AFunctionCaller(iteration, m), writer));
+			
+			ConcurrentUtil.sleep(200);
+			if (count > 0)
+				AFunctionCaller.log(new String(FileUtil.getFileBytes(getFile(iteration).getPath(), null)));
+			AFunctionCaller.log(p + count + " expectations written into '" + getFile(iteration) + p);
+		} finally {
+			FileUtil.close(writer, true);
+		}
 	}
 
 	static File getFile(int iteration) {
 		return FileUtil.userDirFile(fileName + iteration + ".txt");
 	}
 
-	private static void writeExpectation(AFunctionCaller f) {
+	private static void writeExpectation(AFunctionCaller f, BufferedWriter writer) {
+		Thread.currentThread().setUncaughtExceptionHandler(uncaughtExceptionHandler );
 		String then = null;
 		try {
-			if ((def("filter.complextypes", false) && !FunctionCheck.hasSimpleTypes(f)) 
-					|| f.source.getParameterCount() == 0 || void.class.isAssignableFrom(f.source.getReturnType())) {
+			if ((f.source.getParameterCount() == 0 && Modifier.isStatic(f.source.getModifiers())) || void.class.isAssignableFrom(f.source.getReturnType())
+					|| (def("filter.complextypes", false) && !FunctionCheck.hasSimpleTypes(f))) {
 				writeFilteredFunctionCall(f);
 				filter_complextypes++;
 				return;
@@ -172,14 +229,16 @@ public class AutoTestGenerator {
 			return;
 		}
 		count++;
-		FileUtil.writeBytes(expect.getBytes(), getFile(f.cloneIndex).getPath(), true);
+		Util.trY(() -> writer.append(expect));
+//		FileUtil.writeBytes(expect.getBytes(), getFile(f.cloneIndex).getPath(), true);
 	}
 
 	private static void writeFilteredFunctionCall(AFunctionCaller f) {
 		writeFilteredFunctionCall(f.toString());
 	}
 	private static void writeFilteredFunctionCall(String call) {
-		FileUtil.writeBytes((call + "\n").getBytes(), fileName + "filtered.txt", true);
+		Util.trY(() -> filteredFunctionWriter.append(call));
+//		FileUtil.writeBytes((call + "\n").getBytes(), fileName + "filtered.txt", true);
 	}
 	public static Collection<ExpectationFunctionTester> readExpectations(int iteration) {
 		return readExpectations(iteration, getFile(iteration));
@@ -239,26 +298,59 @@ public class AutoTestGenerator {
 class ExpectationCreator {
 	private static final String PREF_WHEN = "@" + Expectations.class.getSimpleName() + "({@" + Expect.class.getSimpleName() + "( when = {";
 	private static final String PREF_THEN = "then = \"";
+	private static final String PREF_CONSTRUCT = "construct = \"";
+	private static final String PREF_CONSTRUCT_TYPES = "constructTypes = \"";
 	
 	static String createExpectationString(AFunctionCaller f, String then) {
 		String expect = "\n@" + Expectations.class.getSimpleName() + "({@" + Expect.class.getSimpleName() 
 				+ "( when = " + Util.toJson(f.getParameter()) 
-				+ " then = \"" + (then != null || f.getResult() == null ? then : (ObjectUtil.isSingleValueType(f.getResult().getClass()) ? FormatUtil.format(f.getResult()) : Util.toJson(f.getResult()))) + "\"})\n"
+				+ " then = \"" + (then != null || f.getResult() == null ? then : (ObjectUtil.isSingleValueType(f.getResult().getClass()) ? FormatUtil.format(f.getResult()) : Util.toJson(f.getResult())))
+				+ (f.construction != null && f.construction.parameter != null ? 
+						" constructTypes = " + Util.toJson(f.getConstruction().constructor.getParameterTypes())
+						+ " construct = " + Util.toJson(f.getConstruction().parameter) : "") 
+				+ "\"})\n"
 				+ f.source + "\n\n";
 		expect = expect.replace("]}", "}");
 		return expect;
 	}
 
-	static Expectations createExpectationFromLine(String l) {
-		String when[], whenall, then;
-		whenall = StringUtil.substring(l, PREF_WHEN, "}");
-		when = whenall.replaceAll("\"", "").split(",");
-		then = StringUtil.substring(l, PREF_THEN, "\"", 0, true);
-		return ExpectationCreator.createExpectation(when, then);
+	private static Class[] getTypes(Object[] construct) {
+		Class[] cs = new Class[construct.length];
+		for (int i = 0; i < construct.length; i++) {
+			cs[i] = construct[i].getClass();
+		}
+		return cs;
 	}
 
-	static ExpectationsImpl createExpectation(String[] when, String then) {
-		return new ExpectationsImpl(new ExpectImpl(-1, null, when, StringUtil.toString(then), -1));
+	static Expectations createExpectationFromLine(String l) {
+		String when[], then, construct[];
+		when = extractArray(l, PREF_WHEN);
+		then = StringUtil.substring(l, PREF_THEN, "\"", 0, true);
+		Class[] constructTypes = loadClasses(extractArray(l, PREF_CONSTRUCT_TYPES));
+		construct = extractArray(l, PREF_CONSTRUCT);
+		return ExpectationCreator.createExpectation(when, then, constructTypes, construct);
+	}
+
+	private static Class[] loadClasses(String[] typenames) {
+		if (typenames == null)
+			return null;
+		Class[] types = new Class[typenames.length];
+		for (int i = 0; i < typenames.length; i++) {
+			types[i] = BeanClass.load(typenames[i]);
+		}
+		return types;
+	}
+
+	private static String[] extractArray(String l, String prefix) {
+		String arr[] = null, all;
+		all = StringUtil.substring(l, prefix, "}", 0, true);
+		if (all != null)
+			arr = all.replaceAll("\"", "").split(",");
+		return arr;
+	}
+
+	static ExpectationsImpl createExpectation(String[] when, String then, Class[] constructTypes, String[] construct) {
+		return new ExpectationsImpl(new ExpectImpl(-1, null, when, StringUtil.toString(then), -1, constructTypes, construct));
 	}
 
 	@SuppressWarnings("unchecked")
@@ -370,14 +462,19 @@ class ExpectImpl implements Expect {
 	private String[] when;
 	private String then;
 	private int resultIndex;
+	private String[] construct;
+	private Class[] constructTypes;
 
-	public ExpectImpl(int parIndex, String whenPar, String[] when, String then, int resultIndex) {
+	@SuppressWarnings("rawtypes")
+	public ExpectImpl(int parIndex, String whenPar, String[] when, String then, int resultIndex, Class[] constructTypes, String[] construct) {
 		super();
 		this.parIndex = parIndex;
 		this.whenPar = whenPar;
 		this.when = when;
 		this.then = then;
 		this.resultIndex = resultIndex;
+		this.constructTypes = constructTypes;
+		this.construct = construct;
 	}
 
 	@Override
@@ -410,6 +507,14 @@ class ExpectImpl implements Expect {
 		return resultIndex;
 	}
 
+	@Override
+	public Class[] constructTypes() {
+		return constructTypes;
+	}
+	@Override
+	public String[] construct() {
+		return construct;
+	}
 	@Override
 	public String toString() {
 		return Util.toJson(this);
